refactor: rewrite README for concision, abandon reviews plan
Reading the new API documentation:

  https://developer.plex.tv/pms/

there doesn't appear to be user reviews exportable. This plan was
hatched when I remembered I once saw user review and summaries show up
in Plex, and I thought that was a feature.

The README has been consolidated, as the code is much more stable and
doesn't need to be comprehensive on its guts as much as it was in the
early hours.