chore(standardization): add python and YAML linters
This commit adds linters for Python via flake8 and black, adds YAML
linting via yamllint, and applies the standardized rules to the project.

As far as app code goes, nothing is structureally changed in this commit
save for a slight clarification in the example systemd unit file, which
uses the local python executable as opposed to the system one in
`/usr/bin/python3`.